KCIRB-1270  - When checking if a parent node is active use the check box instead of the hidden field.  The hidden field represents the persistent state in the database and doesn't change until save when a user alters the active indicator.  - Don't try to delete a node if it's not found to be persistent.

+++ src/main/java/org/kuali/kra/service/impl/ResearchAreasServiceImpl.java
@@ -165,9 +165,11 @@ public class ResearchAreasServiceImpl implements ResearchAreasService {
             // delete before create to allow a research area code to be deleted and reused for a new entry.
             if (details.containsKey("RaDelete")) {
                 ResearchArea researchArea = businessObjectService.findBySinglePrimaryKey(ResearchArea.class, details.get("RaDelete").get("Code"));
-                businessObjectService.delete(researchArea);
-                deleteChildrenResearchAreas(details.get("RaDelete").get("Code"));
-                updateHasChildrenFlag(researchArea.getParentResearchAreaCode());
+                if (researchArea != null) {
+                    businessObjectService.delete(researchArea);
+                    deleteChildrenResearchAreas(details.get("RaDelete").get("Code"));
+                    updateHasChildrenFlag(researchArea.getParentResearchAreaCode());
+                }
             }
             if (details.containsKey("RaCreate")) {
                 boolean active = StringUtils.equalsIgnoreCase(details.get("RaCreate").get("Active"), "true") ? true : false;
